requirements.txt
py-clob-client
websockets
python-dotenv
supabase
